pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "insurance-holder"
authors = [
  { name="Firefox2100", email="wangyunze16@gmail.com" },
]
description = "A project working as an insurance against involuntary disappearance."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "aiomysql~=0.3.2",
    "aiosqlite~=0.22.1",
    "alembic~=1.18.4",
    "argon2-cffi~=25.1.0",
    "asgi-csrf~=0.11",
    "celery~=5.6.2",
    "databases~=0.9.0",
    "fastapi~=0.135.1",
    "httpx~=0.28.1",
    "jinja2~=3.1.6",
    "pyaml~=26.2.1",
    "pydantic~=2.12.5",
    "pydantic-settings~=2.13.1",
    "pymysql~=1.1.2",
    "pytheus~=0.6.1",
    "redis~=7.3.0",
    "sqlalchemy~=2.0.48",
    "uvicorn~=0.42.0",
]
dynamic = [
    "version"
]

[project.urls]
Homepage = "https://github.com/Firefox2100/insurance-holder"
Repository = "https://github.com/Firefox2100/insurance-holder.git"
Issues = "https://github.com/Firefox2100/insurance-holder/issues"

[tool.setuptools.dynamic]
version = {attr = "insurance_holder.__version__"}
